refactor: create a reusable function
